[MDH-107] refactor : redisConfig common 모듈로 이동, 개행 제거, feignConfig 파일 생성, reservation 서비스 코드 좌석 두개 추가 되는 부분 수정
