+++ alarm/src/main/java/com/mdh/alarm/slack/FeignConfig.java
@@ -1,0 +1,9 @@
+package com.mdh.alarm.slack;
+
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.context.annotation.Configuration;
+
+@EnableFeignClients
+@Configuration
+public class FeignConfig {
+}
