+++ alarm/src/test/java/com/mdh/alarm/SlackFallbackTest.java
@@ -6,6 +6,7 @@ import static org.mockito.Mockito.mock;
 
 import com.mdh.alarm.slack.SlackAlertRequest;
 import com.mdh.alarm.slack.SlackClient;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
@@ -19,8 +20,7 @@ class SlackFallbackTest {
     @Mock
     private Throwable mockThrowable;
 
-
-
+    @DisplayName("슬랙 api가 예외를 발생하면 fallback에 작성해놓은 메소드가 호출된다.")
     @Test
     void shouldReturnLoadFailedMessage() {
         // Given
